Relocated meta.js to the components subdir, closer to where it is used and out of the project root.
